스낵 접근 권한 기능 구현, 모임 다이얼로그 수정

See merge request s12-webmobile4-sub1/S12P11D211!148

+++ Witch/app/src/main/java/com/ssafy/witch/data/model/dto/AppointmentDetailItem.kt
@@ -11,8 +11,8 @@ data class AppointmentDetailItem(
     var summary: String,
     var appointmentTime: String,
     var address: String,
-    var latitude: BigDecimal,
-    var longitude: BigDecimal,
+    var latitude: Double,
+    var longitude: Double,
     @SerializedName("members") var participants: List<Participants>
 ) {
     data class Participants (
